refactor: Restructure "datasource" package
Create a "remote" sub-package to differentiate between remote and
(future) local data sources
