+++ app/src/main/java/com/haidoan/android/stren/core/datasource/remote/ExercisesFirestoreDataSource.kt
@@ -1,4 +1,4 @@
-package com.haidoan.android.stren.core.datasource
+package com.haidoan.android.stren.core.datasource.remote
 
 import com.google.firebase.firestore.*
 import com.google.firebase.firestore.ktx.firestore
