+++ app/src/main/java/com/haidoan/android/stren/core/datasource/remote/ExercisesRemoteDataSource.kt
@@ -1,4 +1,4 @@
-package com.haidoan.android.stren.core.datasource
+package com.haidoan.android.stren.core.datasource.remote
 
 import com.google.firebase.firestore.Query
 import com.haidoan.android.stren.core.model.Exercise
