+++ app/src/main/java/com/haidoan/android/stren/core/repository/ExercisesRepositoryImpl.kt
@@ -3,7 +3,7 @@ package com.haidoan.android.stren.core.repository
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.PagingData
-import com.haidoan.android.stren.core.datasource.ExercisesRemoteDataSource
+import com.haidoan.android.stren.core.datasource.remote.ExercisesRemoteDataSource
 import com.haidoan.android.stren.core.model.Exercise
 import com.haidoan.android.stren.core.model.ExerciseCategory
 import com.haidoan.android.stren.core.model.ExerciseFilterStandards
