+++ app/src/main/java/com/haidoan/android/stren/core/repository/WorkoutsRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.haidoan.android.stren.core.repository
 
-import com.haidoan.android.stren.core.datasource.WorkoutRemoteDataSource
+import com.haidoan.android.stren.core.datasource.remote.WorkoutRemoteDataSource
 import com.haidoan.android.stren.core.model.Workout
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.catch
